[Bronze V] Title: 큰 수 (BIG), Time: 132 ms, Memory: 14044 KB -BaekjoonHub

+++ 백준/Bronze/14928. 큰 수 （BIG）/큰 수 （BIG）.java
@@ -1,6 +1,6 @@
 class Main {
     public static void main(String[] args) throws Exception {
-        int c, n = 0;
+        long c, n = 0;
         while ((c = System.in.read()) > 47) {
             n = (n * 10 + (c & 15)) % 20000303;
         }
